Retire the worker-reach-into-orchestrator warning
The rule fired on every request a worker forwarded during node execution, to warn that fetching
orchestrator-owned state per call costs a round trip and returns a view that is stale on arrival.
That concern is real. The rule was not measuring it.

Almost everything it fired on is engine machinery the node author never wrote and cannot avoid --
`get_parameter_value` issuing `ListConnectionsForNodeRequest` per parameter is the loudest, and it
must keep forwarding, because connections are authoritative orchestrator state. So the rule
overwhelmingly indicted the engine while presenting itself as author guidance.

Its remediation then told you to ignore it: "if this is an intentional write (e.g. publishing a
parameter value), ignore." Publishing a parameter value goes through `put_event` and never reaches
this code, so the one case it offered as ignorable cannot occur. A rule that fires constantly on
unavoidable machinery and hands you an excuse teaches people to ignore the whole category, which
costs more than the round trips it was pointing at.

Requests remain the sanctioned boundary -- the engine forbids manager access during worker execution
and directs authors here -- so the guidance moves to the docs, stated once, where it is not
competing with a warning that cries wolf. Scoping the rule to library call sites was the
alternative; it needs a stack walk on every forwarded request during execution, paying a hot-path
cost to warn about hot-path costs.

The strict-mode docs said six rules; the table already listed five, and only the sentence above it
still counted the retired one.

Forwarding behavior is unchanged. The test file that rode along with the warning is renamed to what
it actually covers -- when RemoteHandler forwards versus delegates locally -- which is real and
still tested.
