+++ src/griptape_nodes/app/worker_routing.py
@@ -32,8 +32,6 @@ from dataclasses import dataclass
 from dataclasses import fields as dc_fields
 from typing import TYPE_CHECKING, Any, cast
 
-from griptape_nodes.common.strict_mode import STRICT_MODE
-from griptape_nodes.common.strict_mode_checks import RULES
 from griptape_nodes.retained_mode.events import artifact_events, os_events
 from griptape_nodes.retained_mode.events.base_events import (
     RequestPayload,
@@ -387,11 +385,6 @@ class RemoteHandler:
 
     async def __call__(self, request: RequestPayload) -> ResultPayload:
         if self.event_manager.in_node_execution():
-            rule = RULES["worker-reach-into-orchestrator"]
-            STRICT_MODE.report(
-                rule_id=rule.rule_id,
-                message=rule.render(request_type=type(request).__name__),
-            )
             event_result = await self.event_manager.forward_to_orchestrator(request, ResultContext())
             return cast("ResultPayload", event_result.result)
         return await call_function(self.original, request)
